**Feat: Add project model and related models, enabling display in admin interface**
Created the following models:
  - `ProjectModel`
  - `FrontEndLanguageModel`
  - `BackendLanguageModel`
  - `FrameworkModel`
  - `LibraryModel`

Separated `FrontEndLanguageModel`, `BackendLanguageModel`, `LibraryModel` and `FrameworkModel` into individual models
instead of using `ChoiceField` with predefined options. These models are now linked to `ProjectModel` via a `ManyToManyField`.
This approach allows greater flexibility when adding new front-end languages, back-end languages, frameworks, or libraries. For example,
if a new front-end language is introduced, it can be added directly through the admin interface, rather than modifying the code to update the choices.

Walkthrough:
Instead of hardcoding front-end languages, back-end languages, or frameworks as fixed choices in the code, they can now be dynamically
added in their respective models (FrontEndLanguageModel, BackendLanguageModel, etc.) via the Django admin interface.
Once a new language or framework is added, it will automatically appear as a selectable option in the ProjectModel form without any code changes.

For example, if a new front-end language is added in the FrontEndLanguageModel via the admin interface,
it will then show up as an available option in ProjectModel under front-end languages which in turn enhances the flexibility of the system.

Also updated the data flow to reflect these changes.

- Created the Admin interface to show the models
- Created a superuser to be able to use the admin interface